feat(ha): support ha features (#3478)
* feat(ha): remove sqlite dependency, complete xnode apis

* api: list_agent/list_task

* fix check valid api

* fix metrics send

* fix metrics and agents api

* add ha types

* add rpc client

* fix response action

* fix metrics problems

* add samples, test complete

* remove lazy_static and fs2 deps, use std

* use workspace edition, add xnoded

* init xnoded crate

* create xnode api

* add heartbeat

* complete monitor taosconn

* start_task_job

* task_status

* fix

* api finish

* remove activities

* activities notify

* add response log

* fix log, test apis

* add drop connection action

* fix delete xnode

* fix controller init

* add init logs

* add task status

* fix oneshot

* ignore create xnode error, add main log

* add mnode

* fix mnode role time

* ignore if no leader

* fix monitor

* simplify api methods

* log request extract errors

* fix

* fix

* fix

* fix kafka

* add status log

* fix

* add parser log

* fix

* use unix socket

* fix listener

* fix qid conflict

* log user_pass

* improve alloc method

* fix status

* fix fmt errors

* fix fmt errors

* fix test compile errors

* fix typos

* close e2e tests

* add job logs

* fix concurrency

* fix

* fix

* fix unit test errors

* add more logs

* add logs

* log info

* log level

* add drop task

* add detail log

* improve log

* listener log

* log sql

* fix

* add rebalancer

* fix async file lock and add logs

* add check valid log

* fix job status deserialize

* fix lock file dir not exists error

* fix test

* remove redundant stopping status

* fix http log

* stop job before plan new jobs

* improve event alter status

* fix status update

* add logs

* add log

* remove debug log

* set task status log

* fix multi same event

* fix

* drain xnode

* fix drain

* del job when drop jobs

* fix

* improve rebalance failed task

* improve event codes

* fix drop task

* fix

* fix job status not set error

* filter task id when start job

* fix rebalance auto

* fix rebalance

* start job after reconnect

* fix rebalancer

* fix schedule

* enable rebalance when job start failed

* add log

* fix rebalance manually

* add hb at server side

* add stop log

* add log

* add list task status api

* need to update task status

* fix init

* fix task status alter

* fix cache path

* fix

+++ taosx-core/src/utils/files.rs
@@ -3,10 +3,10 @@ use std::fs::{DirEntry, File};
 use std::io::{Read, Write};
 use std::path::{Path, PathBuf};
 
-use anyhow::{anyhow, Context};
+use anyhow::{Context, anyhow};
 use arrow::array::RecordBatch;
-use base64::engine::general_purpose;
 use base64::Engine;
+use base64::engine::general_purpose;
 use chardetng::EncodingDetector;
 use chrono::{Duration, Utc};
 use encoding_rs::Encoding;
@@ -129,10 +129,10 @@ pub fn decompress_and_write_file(
 pub fn write_to_file(task_id: i64, filename: &String, record: &String) -> anyhow::Result<()> {
     let data_dir = get_data_dir();
     let path = data_dir.join("tasks").join(format!("{task_id}"));
-    if !path.exists() {
-        if let Err(err) = std::fs::create_dir_all(&path) {
-            tracing::error!("failed to create dir {:?}: {}", path, err);
-        }
+    if !path.exists()
+        && let Err(err) = std::fs::create_dir_all(&path)
+    {
+        tracing::error!("failed to create dir {:?}: {}", path, err);
     }
     let path = path.join(format!("{}.{}", filename, Utc::now().format("%Y%m%d")));
     let mut file = OpenOptions::new().create(true).append(true).open(path)?;
@@ -153,10 +153,10 @@ pub fn write_to_parquet_file(
 ) -> anyhow::Result<()> {
     let data_dir = get_data_dir();
     let path = data_dir.join("tasks").join(format!("{task_id}"));
-    if !path.exists() {
-        if let Err(err) = std::fs::create_dir_all(&path) {
-            tracing::error!("failed to create dir {:?}: {}", path, err);
-        }
+    if !path.exists()
+        && let Err(err) = std::fs::create_dir_all(&path)
+    {
+        tracing::error!("failed to create dir {:?}: {}", path, err);
     }
     // delete old files
     if keep_days > 0 {
@@ -191,10 +191,10 @@ pub fn delete_old_parquet_files_by_date(
     let path = data_dir.join("tasks").join(format!("{task_id}"));
     let path = path.join(format!("{}.{}", filename, Utc::now().format("%Y%m%d")));
     let path = path.parent().context("get parent path error")?;
-    if !path.exists() {
-        if let Err(err) = std::fs::create_dir_all(path) {
-            tracing::error!("failed to create dir {:?}: {}", path, err);
-        }
+    if !path.exists()
+        && let Err(err) = std::fs::create_dir_all(path)
+    {
+        tracing::error!("failed to create dir {:?}: {}", path, err);
     }
 
     let cutoff_date = Utc::now() - Duration::days(keep_days as i64);
@@ -206,11 +206,11 @@ pub fn delete_old_parquet_files_by_date(
                 .split('.')
                 .next_back()
                 .context("get file date error")?;
-            if let Ok(file_date) = chrono::NaiveDate::parse_from_str(file_date, "%Y%m%d") {
-                if file_date <= cutoff_date.naive_utc().date() {
-                    tracing::info!("delete archived file: {:?}, since out of date", file);
-                    fs::remove_file(file)?;
-                }
+            if let Ok(file_date) = chrono::NaiveDate::parse_from_str(file_date, "%Y%m%d")
+                && file_date <= cutoff_date.naive_utc().date()
+            {
+                tracing::info!("delete archived file: {:?}, since out of date", file);
+                fs::remove_file(file)?;
             }
         }
     }
@@ -310,10 +310,10 @@ fn read_parquet_dir_entries(task_id: i64, filename: &String) -> anyhow::Result<V
     let path = data_dir.join("tasks").join(format!("{task_id}"));
     let path = path.join(format!("{}.{}", filename, Utc::now().format("%Y%m%d")));
     let path = path.parent().context("get parent path error")?;
-    if !path.exists() {
-        if let Err(err) = std::fs::create_dir_all(path) {
-            tracing::error!("failed to create dir {:?}: {}", path, err);
-        }
+    if !path.exists()
+        && let Err(err) = std::fs::create_dir_all(path)
+    {
+        tracing::error!("failed to create dir {:?}: {}", path, err);
     }
 
     let entries = fs::read_dir(path)?
